About Me
I aim to have a strong foundational understanding of HTML, CSS, and JavaScript. My goal is to be able to build and deploy a basic website independently. Additionally, I plan to acquire essential skills in version control using Git and GitHub. Ultimately, I want to create a portfolio of at least three functional web projects to showcase my abilities to potential employers and be well-prepared for a successful job search in web development.